wip summary chart
